+semver:major Moved the versions of CanWriteToDirectories and CanWriteToDirectory that do not require Windows Forms into SIL.Core
